requirements.txt
pandas
numpy
scikit-learn
tqdm
seaborn
matplotlib
scipy
python-dotenv
openai
notebook
python=3.12
shapely
geopandas
rasterio
rasterstats
shap
xgboost